complete all sections

+++ src/main/java/ir/ac/kntu/logic/Calculator.java
@@ -12,16 +12,16 @@ public class Calculator {
      * @param operator operator to apply on operands.
      * @return result of the calculation
      */
-    public static double calculate(StringBuilder operand1, StringBuilder operand2, StringBuilder operator) {
+    public static Integer calculate(StringBuilder operand1, StringBuilder operand2, StringBuilder operator) {
         switch (Operator.valueOf(operator.toString())) {
             case PLUS:
-                return Double.parseDouble(operand1.toString()) + Double.parseDouble(operand2.toString());
+                return Integer.parseInt(operand1.toString()) + Integer.parseInt(operand2.toString());
             case MINUS:
-                return Double.parseDouble(operand1.toString()) - Double.parseDouble(operand2.toString());
+                return Integer.parseInt(operand1.toString()) - Integer.parseInt(operand2.toString());
             case DIVIDE:
-                return Double.parseDouble(operand1.toString()) / Double.parseDouble(operand2.toString());
+                return Integer.parseInt(operand1.toString()) / Integer.parseInt(operand2.toString());
             case MULTIPLY:
-                return Double.parseDouble(operand1.toString()) * Double.parseDouble(operand2.toString());
+                return Integer.parseInt(operand1.toString()) * Integer.parseInt(operand2.toString());
             case UNKNOWN:
             default:
                 return 0;
